backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.31.0

# Voice
SpeechRecognition>=3.10.0
openai-whisper>=20231117
pyaudio>=0.2.14
pygame>=2.5.0

# ElevenLabs
elevenlabs>=1.0.0

# Tools
pyautogui>=0.9.54
psutil>=5.9.0
pyperclip>=1.8.2
opencv-python>=4.8.0
Pillow>=10.0.0
mss>=9.0.1

# Security
cryptography>=41.0.0

# Optional
spotipy>=2.23.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
